Move path utils to dedication object
