+++ collect_app/src/main/java/org/odk/collect/android/utilities/FormUtils.java
@@ -20,7 +20,7 @@ public class FormUtils {
         String formMediaPath = form.getFormMediaPath();
         return formMediaPath == null
                 ? new ArrayList<>()
-                : FileUtils.listFiles(FileUtils.getFileAtPath(formMediaPath));
+                : FileUtils.listFiles(new File(formMediaPath));
     }
 
     /**
